Remove invoiceninja microvm that used podman rootless
It never worked very well because podman needed a LOT of storage/scratch
space for its images (20+GB). I did not want to pre-allocate that much
space with a disk volume, so tried to use virtiofs shares,
but that didn't work because podman cannot do pivot root on a virtiofs mount.

I will reserve microvm guests for non-containerized applications.